+++ docs/CHANGELOG.md
@@ -1,7 +1,11 @@
+## [v0.12.1]
+- WIP 
+
 ## [v0.12.0]
 - Refactored output director of L2 composite images products 
 - Added `phenocams.group_records_by_unique_filepath`
 - Added `station.get_filtered_fields_as_dict` useful for retriving the L2 & L3 products that are not empty.
+- depreciated `station.generate_query_dict`
   
 ## [v0.11.20]
 - Improved `phenocams.rois_mask_and_sum` to return the sum of pixel values for each channel and not treated as a single channel input image.